Make routing problems non-fatal
In rare circumstances (e.g. due to very bad luck during the map
generation) the autopilot will fail to find a route to the desired
destination. Instead of smashing the stack or killing the gameserver the
lua code returns an error and leaves the problem for the player code to
handle (e.g. by flying elsewhere).