Split SDLview and clientside JSON-Parser
Most of the code in the SDL viewer was actually not related to plotting.
All that code is now moved to ClientLib and can be reused by other
client side programms. A simple such programm is supplied in
StatusBoard. That StatusBoard creates a simple HTML overview of the
current game state.